Updated the GUI to evolve images.
To try out, run either with the command

./run

or run with a selected image like

./run main.py "face.jpg"

or whatever your image name is. I've begrudgingly (but not really)
supplied my face for testing. Feel free to use your own (or someone
else's).

This will allow you to evolve a face from scratch and it will save
the population when you exit the application. However, there is no
way (yet) to save intermediate populations nor is there a way to
reload a past set. (When you see the initial distortions you'll
understand why this wasn't a priority. ;))

Todo:
- MAJOR: Distortions look terrible. The effects are too large. We need
  a way of minimizing the effect of the network on the image.
- Minor: Perhaps a better way of attributing reward rather than winner
  take all?
- Minor: Ability to save/restore populations.